samples: drivers: adc: Increase sample buffer size
Modifies the adc_dt sample to increase the sample buffer size to support
ADCs with more than 16 bits of resolution, such as ad7124.

+++ samples/drivers/adc/adc_dt/src/main.c
@@ -33,7 +33,7 @@ int main(void)
 {
 	int err;
 	uint32_t count = 0;
-	uint16_t buf;
+	uint32_t buf;
 	struct adc_sequence sequence = {
 		.buffer = &buf,
 		/* buffer size in bytes, not number of samples */
